evaluate {evaluate}	R Documentation

Evaluate input and return all details of evaluation.

Description

Compare to eval, evaluate captures all of the information necessary to recreate the output as if you had copied and pasted the code into a R terminal. It captures messages, warnings, errors and output, all correctly interleaved in the order in which they occured. It stores the final result, whether or not it should be visible, and the contents of the current graphics device.

Usage

evaluate(input, envir = parent.frame(), enclos = NULL, debug = FALSE, 
    stop_on_error = 0L, keep_warning = TRUE, keep_message = TRUE, new_device = TRUE, 
    output_handler = default_output_handler)

Arguments

input
input object to be parsed an evaluated. Maybe a string, file connection or function.

envir
environment in which to evaluate expressions

enclos
when envir is a list or data frame, this is treated as the parent environment to envir.

debug
if TRUE, displays information useful for debugging, including all output that evaluate captures

stop_on_error
if 2, evaluation will stop on first error and you will get no results back. If 1, evaluation will stop on first error, but you will get back all results up to that point. If 0 will continue running all code, just as if you'd pasted the code into the command line.

keep_warning,keep_message
whether to record warnings and messages

new_device
if TRUE, will open a new graphics device and automatically close it after completion. This prevents evaluation from interfering with your existing graphics environment.

output_handler
an instance of output_handler that processes the output from the evaluation. The default simply prints the visible return values.

[Package evaluate version 0.5.1 Index]