evaluate-package {evaluate}	R Documentation

evaluate: Parsing and Evaluation Tools that Provide More Details than the Default

Description

Parsing and evaluation tools that make it easy to recreate the command line behaviour of R.

Author(s)

Maintainer: Hadley Wickham hadley@posit.co

Authors:

Yihui Xie (ORCID)

Other contributors:

Michael Lawrence [contributor]

Thomas Kluyver [contributor]

Jeroen Ooms [contributor]

Barret Schloerke [contributor]

Adam Ryczkowski [contributor]

Hiroaki Yutani [contributor]

Michel Lang [contributor]

Karolis Koncevičius [contributor]

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://evaluate.r-lib.org/

https://github.com/r-lib/evaluate

Report bugs at https://github.com/r-lib/evaluate/issues

[Package evaluate version 1.0.5 Index]