inject_funs {evaluate}	R Documentation

Inject functions into the environment of evaluate()

Description

Create functions in the environment specified in the envir argument of evaluate(). This can be helpful if you want to substitute certain functions when evaluating the code. To make sure it does not wipe out existing functions in the environment, only functions that do not exist in the environment are injected.

Usage

inject_funs(...)

Arguments

...
Named arguments of functions. If empty, previously injected functions will be emptied.

Note

For expert use only. Do not use it unless you clearly understand it.

Examples

library(evaluate)
# normally you cannot capture the output of system
evaluate("system('R --version')")

# replace the system() function
inject_funs(system = function(...) cat(base::system(..., intern = TRUE), sep = "\n"))

evaluate("system('R --version')")

inject_funs()  # empty previously injected functions

[Package evaluate version 0.14 Index]