parse_all {evaluate}	R Documentation

Parse, retaining comments.

Description

Works very similarly to parse, but also keeps original formatting and comments.

Usage

parse_all(x, filename = NULL, allow_error = FALSE)

Arguments

x
object to parse. Can be a string, a file connection, or a function. If a connection, will be opened and closed only if it was closed initially.

filename
string overriding the file name

allow_error
whether to allow syntax errors in x

Value

A data.frame with columns src, the source code, and expr. If there are syntax errors in x and allow_error = TRUE, the data frame has an attribute PARSE_ERROR that stores the error object.

[Package evaluate version 0.14 Index]