Parse, retaining comments

Works very similarly to parse, but also keeps original formatting and comments.

Usage

parse_all(x, filename = NULL, allow_error = FALSE)

Arguments

x

object to parse. Can be a string, a file connection, or a function. If a connection, will be opened and closed only if it was closed initially.

filename

string overriding the file name

allow_error

whether to allow syntax errors in x

Value

A data frame with columns src, a character vector of source code, and expr, a list-column of parsed expressions. There will be one row for each top-level expression in x. A top-level expression is a complete expression which would trigger execution if typed at the console. The expression

object in expr can be of any length: it will be 0 if the top-level expression contains only whitespace and/or comments; 1 if the top-level expression is a single scalar (like TRUE, 1, or "x"), name, or call; or 2 if the top-level expression uses ; to put multiple expressions on one line.

If there are syntax errors in x and allow_error = TRUE, the data frame will have an attribute PARSE_ERROR that stores the error object.

Examples

source <- "
  # a comment
  x
  x;y
"
parsed <- parse_all(source)
lengths(parsed$expr)
#> [1] 0 0 1 2
str(parsed$expr)
#> List of 4
#>  $ :  expression()
#>  $ :  expression()
#>  $ :length 1 expression(x)
#>   ..- attr(*, "srcref")=List of 1
#>   .. ..$ : 'srcref' int [1:8] 3 3 3 3 3 3 3 3
#>   .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x55df7be6a7d8> 
#>  $ :length 2 expression(x, y)
#>   ..- attr(*, "srcref")=List of 2
#>   .. ..$ : 'srcref' int [1:8] 4 3 4 3 3 3 4 4
#>   .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x55df7be6a7d8> 
#>   .. ..$ : 'srcref' int [1:8] 4 5 4 5 5 5 4 4
#>   .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x55df7be6a7d8> 
#>  - attr(*, "class")= chr "AsIs"