parse_all {evaluate}	R Documentation

Parse, retaining comments.

Description

Works very similarly to parse, but also keeps original formatting and comments.

Usage

parse_all(x)

Arguments

x
object to parse. Can be a string, a file connection, or a function

Value

a data.frame with columns src, the source code, and eval

[Package evaluate version 0.5.1 Index]