remove_hooks {evaluate}	R Documentation

Remove hooks.

Description

This provides a way to remove previously set hook values.

Usage

remove_hooks(hooks)

Arguments

hooks
a named list of hooks - each hook can either be a function or a list of functions.

Examples

new1 <- list(before.plot.new = function() print("Plotted!"))
new2 <- list(before.plot.new = function() print("Plotted Again!"))
set_hooks(new1)
set_hooks(new2)
plot(1)
remove_hooks(new1)
plot(1)
remove_hooks(new2)
plot(1)

[Package evaluate version 0.14 Index]